pyproject.toml
[project]
name = "aks"
version = "0.1.0"
description = "Agent Knowledge System — personal AI assistant grounded in your notes"
requires-python = ">=3.12"
license = { text = "MIT" }
dependencies = [
    "anthropic>=0.40",
    "click>=8.0",
    "pyyaml>=6.0",
    "python-dotenv>=1.0",
]

[project.scripts]
aks = "aks.main:cli"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/aks"]

[tool.uv]
dev-dependencies = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
